Update README.md
I added explanation of how the prediction of error (-0.1333) is done for the fibo sequence.

I removed the Conclusion section because it was unclear in purpose and the expression there at the end is incorrect. 
As for showing an example that the method produces the correct prediction - it was already done earlier in the Readme.